Create the project
The project is a test for Aubay. The goals are clear : fetch rss feed from the internet and display them into a list. When the user click on one of the article, a new view with details must be displaying. It's possible to click on a button to see all the article inside a web view. The design should be clean, it's possible to get some inspiration thanks to apps like Feedly.