Configure injection, Rx java, ViewModel
This step is the next stage of the MvvM configuration with retrofit, dagger and RxJava. To make simple I handled the api instance injection with Dagger, then the required dependencies can be injected in BaseViewModel and Thanks to RxJava coupled with injection I'am able to retrieve data in ViewModel. And in RssFeedListViewModel class we can get the result. And the subscription is also taking care too.