Switch Gson converter to Xml converter
The feeds are in Xml so we have to parse it as Xml object. It's necessary to use Simple Xml because Jaxb doesn't seem to work on Android. The pojo object are updated and Retrofit instance as well. It was necessary to switch the call in order to get the elements in the correct order.